+++ lib/full-site-editing/edit-site-page.php
@@ -49,10 +49,8 @@ function gutenberg_get_editor_styles() {
 		),
 	);
 
-	/* translators: Use this to specify the CSS font family for the default font. */
-	$locale_font_family = '-apple-system, BlinkMacSystemFont,"Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell,"Helvetica Neue", sans-serif';
-	$styles[]           = array(
-		'css' => "body { font-family: '$locale_font_family' }",
+	$styles[] = array(
+		'css' => 'body { font-family: -apple-system, BlinkMacSystemFont,"Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell,"Helvetica Neue", sans-serif }',
 	);
 
 	if ( $editor_styles && current_theme_supports( 'editor-styles' ) ) {