+++ packages/data/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+- Added new `resolveSelect` registry method to initiate and wait for selector resolution
+
 ## 4.26.0 (2020-12-17)
 
 ### New Features